Removed un necessary code from Controller and Model file
